requirements.txt
geopandas
matplotlib
numpy
pre-commit
rasterio
segmentation_models
scikit-learn
jupytext==1.14.5
jupyter
tensorflow[and-cuda]
keras==2.15.0
Path
datetime
google.cloud
pytz
pynvml
numba
opencv-python
folium
segmentation_models
